Se separaron los home para user y para translator, se movieron todos los componentes y archivos de estilos que no son paginas a ´components´
