+++ src/components/tramiteCard/tramiteCard.css
