+++ src/components/translator/translator.css
