+++ src/components/translator/translator.js
